app_files/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
fastapi==0.123.5
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6
numpy==2.3.5
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
setuptools==80.9.0
starlette==0.50.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
python-multipart
